To_Send/requirements.txt
opencv-python
numpy
pandas
scikit-learn
xgboost
tensorflow
mtcnn
joblib
Pillow
pillow-heif
tqdm